chore: Rename the repo to ss-python.
