+++ .github/workflows/readthedocs-preview.yml
@@ -7,7 +7,7 @@ jobs:
     steps:
       - uses: readthedocs/actions/preview@v1
         with:
-          project-slug: serious-scaffold-python
+          project-slug: ss-python
 name: Read the Docs Pull Request Preview
 on:
   pull_request_target:
