+++ includes/variable.jinja
@@ -23,11 +23,11 @@
 [%- endmacro %]
 
 [% macro logo_badge() -%]
-  [![Serious Scaffold Python]({{ logo_badge_url() }})](https://serious-scaffold.github.io/serious-scaffold-python)
+  [![Serious Scaffold Python]({{ logo_badge_url() }})](https://serious-scaffold.github.io/ss-python)
 [%- endmacro %]
 
 [% macro logo_badge_url() -%]
-  https://img.shields.io/endpoint?url=https://serious-scaffold.github.io/serious-scaffold-python/_static/badges/logo.json
+  https://img.shields.io/endpoint?url=https://serious-scaffold.github.io/ss-python/_static/badges/logo.json
 [%- endmacro %]
 
 [% macro page_url() -%]
