+++ test/settings_test.py
@@ -1,13 +1,13 @@
 """Test for settings."""
 import os
 
-from serious_scaffold.settings import global_settings, settings
+from ss_python.settings import global_settings, settings
 
 
 def test_settings() -> None:
     """Test for settings."""
     assert settings.logging_level == os.getenv(
-        "SERIOUS_SCAFFOLD_LOGGING_LEVEL",
+        "SS_PYTHON_LOGGING_LEVEL",
         "INFO",
     )
     assert str(global_settings.ci).lower() == os.getenv("CI", "False").lower()
